Move Resolve to lib/
